Add interfaces abd profile resolver

+++ src/app/pages/menu/menu.component.ts
@@ -15,7 +15,9 @@ export class MenuComponent {
   readonly menuItems = menuItems;
 
   goTo(route: string): void {
-    this.authservice.removeToken();
+    if(route === 'login') {
+      this.authservice.removeToken();
+    }
     this.router.navigateByUrl(route);
   }
 }
